requirements.txt
streamlit
google-generativeai
langchain_google_genai
python-dotenv
faiss-cpu
